lab4/requirements.txt
bcrypt==4.1.3
cffi==1.16.0
cryptography==42.0.7
future==1.0.0
jinja2==3.1.4
MarkupSafe==2.1.5
netmiko==4.3.0
ntc-templates==5.0.0
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
pyserial==3.5
PyYAML==6.0.1
scp==0.14.5
six==1.16.0
textfsm==1.1.3
